Move MrpFeedbackConfig from Types.h to Algorithm.h
